Remove usage of iff function.

+++ score-http/src/main/java/org/oagi/score/gateway/http/api/context_management/data/ContextSchemeValue.java
@@ -3,15 +3,16 @@ package org.oagi.score.gateway.http.api.context_management.data;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 
 @Data
 public class ContextSchemeValue implements Serializable {
 
-    private long ctxSchemeValueId;
+    private BigInteger ctxSchemeValueId;
     private String guid;
     private String value;
     private String meaning;
-    private long ownerCtxSchemeId;
+    private BigInteger ownerCtxSchemeId;
     boolean used;
 
 }
